Re-activate assemble script
Not adding a previously existing component from the files that will be
pushed to the update site will NOT actually REMOVE it from the site, it
will only be removed from the database!

As a result, files can still be downloaded "manually" from the update
site repo, only when using the Fiji Updater they will not be added to an
installation any more (as they're not in the DB).