Start restructuring repo layout
In order for being able to specify which extra files should go to what
update site, they also need to be part of a site-specific layout. This
commit moves the site-specific stuff to the top-level directory, so the
other parts of the repo can be adjusted as well in later commits and it
is immediately visible what parts should go to which site.